Move DevicesSubsystem into Devices.scala; clean up Hierarchy companion object
