+++ src/main/scala/util/Hierarchy.scala
@@ -6,36 +6,14 @@ import chisel3._
 import chisel3.util._
 
 import freechips.rocketchip.config._
-import freechips.rocketchip.diplomaticobjectmodel.logicaltree.LogicalTreeNode
 import freechips.rocketchip.tilelink._
 import freechips.rocketchip.subsystem._
 import freechips.rocketchip.diplomacy._
-import freechips.rocketchip.prci._
 import freechips.rocketchip.util.LocationMap
 
 import firrtl.graph._
 
 case object HierarchyKey extends Field[Option[DiGraph[HierarchicalLocation]]](None)
-
-case class DevicesSubsystemParams(
-  name: String,
-  logicalTreeNode: LogicalTreeNode,
-  asyncClockGroupsNode: ClockGroupEphemeralNode)
-
-class DevicesSubsystem(val location: HierarchicalLocation, val ibus: InterruptBusWrapper, params: DevicesSubsystemParams)(implicit p: Parameters)
-  extends LazyModule 
-    with Attachable
-    with HasConfigurableTLNetworkTopology 
-    with CanHaveDevices {
-
-  def devicesSubhierarchies = None
-  def logicalTreeNode = params.logicalTreeNode
-  implicit val asyncClockGroupsNode = params.asyncClockGroupsNode
-
-  lazy val module = new LazyModuleImp(this) {
-    override def desiredName: String = params.name
-  }
-}
 
 trait HasConfigurableHierarchy { this: Attachable =>
   def location: HierarchicalLocation
@@ -52,8 +30,8 @@ trait HasConfigurableHierarchy { this: Attachable =>
     edges.foreach { edge =>
       val dssParams = DevicesSubsystemParams(
         name = edge.name,
-        logicalTreeNode = this.logicalTreeNode,
-        asyncClockGroupsNode = this.asyncClockGroupsNode)
+        logicalTreeNode = context.logicalTreeNode,
+        asyncClockGroupsNode = context.asyncClockGroupsNode)
       val dss = context { LazyModule(new DevicesSubsystem(edge, ibus, dssParams)) }
       createHierarchyMap(edge, graph, dss)
     }
@@ -67,7 +45,6 @@ trait HasConfigurableHierarchy { this: Attachable =>
       .filter(_.location != location)
   }
 
-  val busLocationFunctions = LocationMap.empty[LocationMap[TLBusWrapper]]
   val hierarchyMap = LocationMap.empty[Attachable]
 
   p(HierarchyKey).foreach(createHierarchyMap(location, _, this))
@@ -77,34 +54,8 @@ trait HasConfigurableHierarchy { this: Attachable =>
   }
 }
 
-class Hierarchy(val root: HierarchicalLocation) {
-  require(root == InSystem || root == InSubsystem, "Invalid root hierarchy")
-
-  val graph = new MutableDiGraph[HierarchicalLocation]
-  graph.addVertex(root)
-
-  def addSubhierarchy(parent: HierarchicalLocation, child: HierarchicalLocation): Unit = {
-    graph.addVertex(child)
-    graph.addEdge(parent,child) 
-  }
-
-  def addSubhierarchies(parent: HierarchicalLocation, children: Seq[HierarchicalLocation]): Unit = {
-    children.foreach(addSubhierarchy(parent,_))
-  }
-
-  def closeHierarchy(): DiGraph[HierarchicalLocation] = {
-    DiGraph(graph)
-  }
-}
-
 object Hierarchy {
-  def init(root: HierarchicalLocation): Hierarchy = {
-    val hierarchy = new Hierarchy(root)
-    hierarchy
-  }
-
-  def default(root: HierarchicalLocation): DiGraph[HierarchicalLocation] = {
-    val h = init(root)
-    h.closeHierarchy()
+  def apply(edges: Map[HierarchicalLocation, Set[HierarchicalLocation]]): DiGraph[HierarchicalLocation] = {
+    DiGraph(edges)
   }
 }
